Fix random number generation
In Perl, int(rand 4) returns a random integer between 0 and 3, inclusive:

  $ perl -le "for (1..1000) { print(int(rand 4)) }" | sort | uniq -c
      248 0
      274 1
      241 2
      237 3

In Python, random.randint(a, b) returns a random integer N such that a <= N <= b:

  $ python3 -c $'import random\nfor i in range(1, 1000): print(random.randint(0, 3))' | sort | uniq -c
      261 0
      247 1
      241 2
      250 3

In Perl, rand returns a random fractional number greater than or equal to 0 and
less than 1. In Python, random.random() is equivalent.

Links:

 * https://perldoc.perl.org/functions/rand.html

 * https://docs.python.org/3/library/random.html